Change the interface of card and adding info package
