+++ SuperCardGame/app/src/main/java/cs446_w2018_group3/supercardgame/info/GameInfo.java
@@ -1,0 +1,9 @@
+package cs446_w2018_group3.supercardgame.info;
+
+/**
+ * Created by yandong on 2018-02-25.
+ */
+
+public class GameInfo {
+
+}
