+++ SuperCardGame/app/src/main/java/cs446_w2018_group3/supercardgame/model/cards/AirCard.java
@@ -1,0 +1,13 @@
+package cs446_w2018_group3.supercardgame.model.cards;
+
+/**
+ * Created by yandong on 2018-02-25.
+ */
+
+public class AirCard extends ElementCard {
+        public AirCard () {
+            this.id = 2;
+            this.label = "Air";
+            this.level = 1;
+        }
+    }
