+++ SuperCardGame/app/src/main/java/cs446_w2018_group3/supercardgame/model/cards/DirtCard.java
@@ -1,0 +1,13 @@
+package cs446_w2018_group3.supercardgame.model.cards;
+
+/**
+ * Created by yandong on 2018-02-25.
+ */
+
+public class DirtCard extends ElementCard {
+        public DirtCard () {
+            this.id = 3;
+            this.label = "Dirt";
+            this.level = 1;
+        }
+    }
