+++ SuperCardGame/app/src/main/java/cs446_w2018_group3/supercardgame/model/cards/ElementCard.java
@@ -1,0 +1,14 @@
+package cs446_w2018_group3.supercardgame.model.cards;
+
+import cs446_w2018_group3.supercardgame.model.cards.Card;
+
+/**
+ * Created by yandong on 2018-02-25.
+ */
+
+public class ElementCard extends Card {
+    int level;
+    int damage;
+    @Override
+    public void apply() {}
+}
