+++ SuperCardGame/app/src/main/java/cs446_w2018_group3/supercardgame/model/cards/FireCard.java
@@ -1,0 +1,13 @@
+package cs446_w2018_group3.supercardgame.model.cards;
+
+/**
+ * Created by yandong on 2018-02-25.
+ */
+
+public class FireCard extends ElementCard {
+        public FireCard () {
+            this.id = 1;
+            this.label = "Fire";
+            this.level = 1;
+        }
+    }
