+++ SuperCardGame/app/src/main/java/cs446_w2018_group3/supercardgame/model/cards/ItemCard.java
@@ -1,0 +1,11 @@
+package cs446_w2018_group3.supercardgame.model.cards;
+
+/**
+ * Created by yandong on 2018-02-25.
+ */
+
+public class ItemCard extends Card {
+    @Override
+    public void apply() {}
+}
+
